[WIP] Prepare base for alternative repair method
Since de-optimization and repair via smali/basmali is broken
we seek alternative methods via ART oat utils. The idea is to start
dumping DEX from OAT rodata section, repair the checksum and
let the dex2oat do the offset recalculation work again when pre-opt
for AOSP build. Have partially being test and not sure that it works
across the border. We'll just give it a shot.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>