Cherrypick individual modules shared libs
Continuing the effort to fix [1], it seems that we can't blindly
upgrade all shared libraries from PRODUCT_COPY_FILES
to individual modules. Most of the supported devices conflict
in multiple places for re-definitions. While most case can be
dealt with some override options some can't.

As such we have to proceed with shared libs cherrypicking
to indicate which libraries from the master blobs list should
be defined as a separate module.

For this purpose the generate-vendor script is taking an
additional argument with a txt file which contains such libs
for each target device. File templates are located under
devices root dir (same with proprietary blobs list) and will
be periodically be updated with the required dependencies.

First target is support for N5x.

[1] https://github.com/anestisb/android-prepare-vendor/issues/21

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>