Fix undefined variable bug
If not shared libraries modules are selected for a
provided partition, PKGS array is not defined resulting
to a runtime error. Add a boolean flag to ensure that array
is iterated only when elements available.

We probably need to re-consider the way we want to use
dynamically appended arrays in bash to avoid such cases,
although for the time being leave it for a later clean-up.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>