Never DEX preopt repaired APKs/JARs
Mitigate issue https://github.com/anestisb/android-prepare-vendor/issues/22#issuecomment-245216672

This effectively means that we instruct the build to not optimize
the repaired bytecode, just zipalign & resign them. As such the
bytecode will be optimized at device first boot. This was the only
option to circumvent "--abort-on-hard-verifier-error" without
modifying the AOSP default makefiles.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>