Split configuration into naked & gapps comatible
This is the base commit of small configuration base refactor so that
we can maintain two versions. The naked one which has all the blobs
required to have a functional device without Google Play Services.
The gplay branch has extra blobs / files which are required only when
GApps are installed.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>