Fix issue #36
We do all the XSRF & NID cookie work manually so that a valid
POST request is made against the ToS accept URL.

I've maintained the ToS message as presented in original website
for obvious reasons. By default user is questioned to accept ToS
prior to start downloading the image.

There is an auto accept (-y|--yes) flag available so that scripting
automation is not broken due to required user interaction from stdin.

The whole situation is total BS, although I try to keep it as formal
as possible with all the sensitive ToS bits.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>