Different approach for flounder dual target boards
Unfortunately volantis & volantisg are not separate targets in
AOSP tree as such there is no official way to diffrentiate at
Makefile level the desired target board version. As such we
leave the responsibility to the user in order to include the
appropriate version of flounder vendor blobs. We just add a
fatal warning in case both Makefile trees are detected under
same AOSP root.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>