leave out CarrierSetup for Pixel naked config
CarrierSetup doesn't appear to have any use in the naked configuration
build. It depends on Verizon code that's not included and is triggered
by configuration in the GoogleCarrierConfig vendor.xml not present in
AOSP. CopperheadOS includes the stock vendor.xml as an overlay, but
leaves out the Verizon configuration since it depends on the app suite
and isn't required for basic LTE+ functionality without VoLTE / VoWiFi.

The permissions whitelist for CarrierSetup is also in the file for
Google Play Services (privapp-permissions-google.xml) rather than the
one for the Pixels (privapp-permissions-marlin.xml). If it's included,
that whitelist needs to be included too.