avoid treating vendor as a radio file
This avoids generating an incorrect add-radio-file call for the vendor
image, causing the stock factory image to be included and flashed by the
factory images and breaking booting. It moves the marlin vendor image
inclusion to the shared AB_OTA_PARTITIONS generation so it gets applied
for taimen / walleye too.